refactor: reduce cognitive complexity
